% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestOutliers.R
\name{Compare}
\alias{Compare}
\title{Compare datafiles}
\usage{
Compare(object, what = "DataRosner")
}
\arguments{
\item{object}{An object of type testOutliers}

\item{what}{A dataframe to compare}
}
\value{
either names or nothing
}
\description{
Compare datafiles
}
