% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestOutliers.R
\name{Compare.testOutlier}
\alias{Compare.testOutlier}
\title{Compare datafiles}
\usage{
\method{Compare}{testOutlier}(object, what = "DataRosner")
}
\arguments{
\item{object}{An object of type testOutliers}

\item{what}{}
}
\value{
either names or nothing
}
\description{
Compare datafiles
}
