% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestOutliers.R
\name{getData.testOutlier}
\alias{getData.testOutlier}
\title{getData}
\usage{
getData.testOutlier(object, data = "?")
}
\arguments{
\item{object}{An object of type testOutliers}

\item{data}{A name of the data that should be returned. If data = '?' the function
will return all the names in the object that can be used for data (All those
names starts with 'DataSomthing')}
}
\value{
a dataframe handled by a specific method
}
\description{
getData
}
