% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestOutliers.R
\name{plot.testOutlier}
\alias{plot.testOutlier}
\title{Plot
This function plots outlier diagnostics for the model}
\usage{
\method{plot}{testOutlier}(x, ...)
}
\arguments{
\item{x}{object of type testOutliers}

\item{...}{Not used now}
}
\value{
Nothing
}
\description{
Plot
This function plots outlier diagnostics for the model
}
