% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestOutliers.R
\name{plotTestOutlier}
\alias{plotTestOutlier}
\title{plotTestOutlier
This function plots outlier diagnostics for the model}
\usage{
plotTestOutlier(x, ...)
}
\arguments{
\item{x}{object of type testOutliers}

\item{...}{Not used now}
}
\value{
Nothing
}
\description{
plotTestOutlier
This function plots outlier diagnostics for the model
}
