% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestOutliers.R
\name{summary.testOutlier}
\alias{summary.testOutlier}
\title{summary}
\usage{
\method{summary}{testOutlier}(object, ...)
}
\arguments{
\item{object}{An object of type testOutliers}

\item{...}{Not used now
This function prints summary of all lm models with fixed outliers}
}
\value{
nothing
}
\description{
summary
}
