% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestOutliers.R
\name{summaryTestOutlier}
\alias{summaryTestOutlier}
\title{summaryTestOutlier}
\usage{
summaryTestOutlier(object, ...)
}
\arguments{
\item{object}{An object of type testOutliers}

\item{...}{Not used now
This function prints summary of all lm models with fixed outliers}
}
\value{
nothing
}
\description{
summaryTestOutlier
}
