% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestOutliers.R
\name{testOutliers}
\alias{testOutliers}
\title{testOutliers
This is the main function for the package, it takes a lm formula
and data and handels outliers in different ways. Thereafter it
recalculates the lm model with data that has been handled}
\usage{
testOutliers(form, data, limit = 0.001, missing = "None", otherVar = c())
}
\arguments{
\item{form}{an lm formula}

\item{data}{data that should include all variables in the lm formula}

\item{limit}{sets alpha that is used for some outlier detection models}

\item{missing}{not used}

\item{otherVar}{not used}
}
\value{
a testOutliers object that include summary of all models
and all datafiles, including the original.
}
\description{
testOutliers
This is the main function for the package, it takes a lm formula
and data and handels outliers in different ways. Thereafter it
recalculates the lm model with data that has been handled
}
